feat: adicionando funcionalidade extras e necessarias em entity, cont…

+++ trabalhoFinalApi/src/main/java/org/serratec/trabalhoFinalApi/service/CategoriaService.java
@@ -2,12 +2,15 @@ package org.serratec.trabalhoFinalApi.service;
 
 import lombok.AllArgsConstructor;
 import org.serratec.trabalhoFinalApi.entity.Categoria;
+import org.serratec.trabalhoFinalApi.exception.catalogo.CategoriaConflitoProdutosException;
+import org.serratec.trabalhoFinalApi.exception.catalogo.CategoriaJaCadastrada;
 import org.serratec.trabalhoFinalApi.exception.catalogo.CategoriaPaiNaoEncontradaException;
 import org.serratec.trabalhoFinalApi.exception.generalista.RequisicaoNaoEncontradoException;
 import org.serratec.trabalhoFinalApi.model.CategoriaDto.CategoriaAtualizar;
 import org.serratec.trabalhoFinalApi.model.CategoriaDto.CategoriaBuscar;
 import org.serratec.trabalhoFinalApi.model.CategoriaDto.CategoriaCriar;
 import org.serratec.trabalhoFinalApi.repository.CategoriaRepository;
+import org.serratec.trabalhoFinalApi.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -19,6 +22,7 @@ import java.util.UUID;
 @Service
 public class CategoriaService {
 
+    private ProdutoRepository produtoRepository;
     private CategoriaRepository categoriaRepository;
 
     public List<CategoriaBuscar> buscarCategoria(UUID id, String nome){
@@ -29,7 +33,7 @@ public class CategoriaService {
         } else if (id != null) {
             Optional<Categoria> categoria = this.categoriaRepository.findById(id);
             if(categoria.isEmpty()){
-                throw new CategoriaPaiNaoEncontradaException(id);
+                throw new RequisicaoNaoEncontradoException("O id ou o nome especificado nao condiz com um existente no banco de dados");
             }
             categorias.add(categoria.get());
         } else if (nome != null && !nome.isBlank()){
@@ -47,6 +51,9 @@ public class CategoriaService {
     }
 
     public CategoriaBuscar cadastrarCategoria(CategoriaCriar categoriaCriar){
+        if(categoriaRepository.existsByNomeIgnoreCase(categoriaCriar.getNome())){
+            throw new CategoriaJaCadastrada(categoriaCriar.getNome());
+        }
         Categoria categoria = new Categoria(categoriaCriar);
         return new CategoriaBuscar(categoriaRepository.save(categoria));
 
@@ -54,7 +61,10 @@ public class CategoriaService {
 
     public void deletarCategoria(UUID id){
         if(!categoriaRepository.existsById(id)){
-            throw new CategoriaPaiNaoEncontradaException(id);
+            throw new RequisicaoNaoEncontradoException("O id especificado nao existe no banco de dados");
+        }
+        if(produtoRepository.existsByCategoriaId(id)){
+            throw new CategoriaConflitoProdutosException(id);
         }
         this.categoriaRepository.deleteById(id);
     }
